Section 2B Graybox
Created the Graybox level for Section 2B. There's a "Section 2B 1-3" scene containing the level, as well as a "Section 2B" prefab of the level. It is currently blocked out with walls and a placeholder terrain. The mechanic elements of the level are blocked out with placeholder items. (Bushes are used for glowstick supply, etc.) They can be replaced later when the mechanics are in.