Feature/gdc (#123)
* GDC test Area 1 & 2

* Minor Changes

* Camera Triggers

* Unselect moveable mirror after certain distance

* Fix camera change trigger

* Area 1 and 2 with fixes

* Level rock

* Trying Things

* Area 3

* Environment Changes

* Area 4

* Area 4 Changes

* Added environment lights, materials etc.

* minor changes

* Runner and Runner(1) jump

* Minor navmesh and value changes

* Merge

* Fix

* Updated Hierarchy and level prefabs

* Section 1 Changes

* Updated Camera Angles

Updated some of the camera angles in the GDC Test scene. Some will still need adjustments as more of the level is added and I believe there are some issues with the camera system we should iron out to make it smoother.

* Added etchings to GDC scene

Added etchings to GDC scene. I also made the ones I used into Prefabs and put them in a folder so they're easier to place next time and you don't have to go through the entire process of putting the sprites together and throwing scripts on them.

* Section 1 Level Changes + Hierarchy (Prefab) Changes + Etching Changes

* Updated Camera Angles

Updated camera angles based on changes to level.

* Level Changes w/ triggers

* Camera quickfix

* Section 1 Level Polish w/ checkpoint

* Section 1 Polish + Narrative System

* Section 1 Update w/ Fog + environment Art

* World Building Polish

* NewMenu

A new iteration of the main menu

* World Building Polish + Navmesh Update+ Level 2 Start

* Minor Changes

* Fix scene

* AIMagic scene added

* NewMenuRevised

- New logo
- Hopefully better lighting (built and checked)
- Lesser foliage
- Fixed the light that follows the selected option
I think it's good enough now.. if you'll want to make any changes, I pushed it to the newmenu branch

* Section 1 Polish after playtest

* Minor Trigger Changes

* Minor Changes

* Prefab Updates + Area 7 Graybox

* Minor Changes + Adding new diary system

* Area 8 + Section 1 Polish