Fixed doors and minor changes
Fixed doors in level with GDC level and ErinSegments integrated. The level still needs polish, and I think it needs pithacks in the areas where the enemies are supposed to jump off. But it should be good to go for Varun to integrate his scene. Checkpoints and camera angles work.