Erin (#203)
* Fixed issues with scene

* Scene changes; Occlusion culling

* Added enemies and checkpoints

* Made Prefab of "ErinSegments", started integrating

Started integrating in new scene "Integrating-GDC-Erin". We shouldn't need to change the camera angles if we don't rotate the level segments.

* Integrated ErinSegments with GDC

ErinSegments is integrated with GDC, with the exception of the doors need to be fixed. Checkpoints and camera angles are working correctly.

* Fixed doors and minor changes

Fixed doors in level with GDC level and ErinSegments integrated. The level still needs polish, and I think it needs pithacks in the areas where the enemies are supposed to jump off. But it should be good to go for Varun to integrate his scene. Checkpoints and camera angles work.

* attempting to fix things to merge into develop